updated old code in case new one is broken

+++ Regolatore.cpp
@@ -6,14 +6,14 @@ using namespace std;
 class Regolatore
 {
 private:
-    vector<double> output_coefficients;
-    vector<double> input_coefficients;
-    vector<double> previous_inputs;
-    vector<double> previous_outputs;
+    vector<float> output_coefficients;
+    vector<float> input_coefficients;
+    vector<float> previous_inputs;
+    vector<float> previous_outputs;
 
 public:
-    Regolatore(vector<double> output_coeff, vector<double> input_coeff);
-    double calculate_output(double input);
+    Regolatore(vector<float> output_coeff, vector<float> input_coeff);
+    float calculate_output(float input);
     void reset();
 };
 
@@ -22,7 +22,7 @@ void Regolatore::reset(){
     previous_outputs.resize(previous_outputs.size(), 0);
 }
 
-double Regolatore::calculate_output(double input)
+float Regolatore::calculate_output(float input)
 {
     // Aggiorna il vettore degli ingressi precedenti spostando tutti i valori di un posto in avanti
     for (int i = previous_inputs.size() - 1; i > 0; i--)
@@ -33,11 +33,11 @@ double Regolatore::calculate_output(double input)
     previous_inputs[0] = input;
     // Il contributo dell'ingresso all'uscita è il prodotto scalare tra il vettore degli ingressi precedenti che comprende quello attuale
     // e i coefficienti relativi agli ingressi della funzione di trasferimento del regolatore
-    double output_input_component = inner_product(previous_inputs.begin(), previous_inputs.end(), input_coefficients.begin(), 0.0);
+    float output_input_component = inner_product(previous_inputs.begin(), previous_inputs.end(), input_coefficients.begin(), 0.0);
 
     // Il contributo dell'uscita precedente all'uscita attuale è il prodotto scalare tra il vettore delle uscite precedenti
     // e i coefficienti relativi all'uscita della funzione di trasferimento del regolatore
-    double output_output_component = inner_product(previous_outputs.begin(), previous_outputs.end(), output_coefficients.begin(), 0.0);
+    float output_output_component = inner_product(previous_outputs.begin(), previous_outputs.end(), output_coefficients.begin(), 0.0);
 
     
     for (int i = previous_outputs.size() - 1; i > 0; i--)
@@ -50,7 +50,7 @@ double Regolatore::calculate_output(double input)
     return previous_outputs[0];
 }
 
-Regolatore::Regolatore(vector<double> output_coeff, vector<double> input_coeff)
+Regolatore::Regolatore(vector<float> output_coeff, vector<float> input_coeff)
 {
     output_coefficients = output_coeff;
     input_coefficients = input_coeff;
